Changed counter from int to long

+++ u2flib-server-core/src/main/java/com/yubico/u2f/data/DeviceRegistration.java
@@ -119,7 +119,7 @@ public class DeviceRegistration extends JsonObject implements Serializable {
         return super.toJson();
     }
 
-    public void checkAndUpdateCounter(int clientCounter) throws U2fException {
+    public void checkAndUpdateCounter(long clientCounter) throws U2fException {
         if (clientCounter <= counter) {
             throw new InvalidDeviceCounterException();
         }
